+++ source/ui/src/main/java/com/clerk/ui/signin/code/SignInFactorCodeView.kt
@@ -13,16 +13,16 @@ import com.clerk.api.Clerk
 import com.clerk.api.network.model.factor.Factor
 import com.clerk.api.ui.ClerkTheme
 import com.clerk.ui.R
+import com.clerk.ui.auth.AuthStateEffects
 import com.clerk.ui.auth.Destination
 import com.clerk.ui.auth.LocalAuthState
 import com.clerk.ui.auth.PreviewAuthStateProvider
+import com.clerk.ui.auth.VerificationUiState
+import com.clerk.ui.auth.verificationState
 import com.clerk.ui.core.button.standard.ClerkTextButton
-import com.clerk.ui.core.common.AuthStateEffects
 import com.clerk.ui.core.common.Spacers
 import com.clerk.ui.core.common.StrategyKeys
-import com.clerk.ui.core.common.VerificationUiState
 import com.clerk.ui.core.common.scaffold.ClerkThemedAuthScaffold
-import com.clerk.ui.core.common.verificationState
 import com.clerk.ui.core.input.ClerkCodeInputField
 import com.clerk.ui.theme.ClerkMaterialTheme
 import com.clerk.ui.theme.DefaultColors
