+++ source/ui/src/main/java/com/clerk/ui/signin/password/reset/SignInSetNewPasswordView.kt
@@ -25,11 +25,11 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.clerk.ui.R
 import com.clerk.ui.auth.AuthState
+import com.clerk.ui.auth.AuthStateEffects
+import com.clerk.ui.auth.AuthenticationViewState
 import com.clerk.ui.auth.LocalAuthState
 import com.clerk.ui.auth.PreviewAuthStateProvider
 import com.clerk.ui.core.button.standard.ClerkButton
-import com.clerk.ui.core.common.AuthStateEffects
-import com.clerk.ui.core.common.AuthenticationViewState
 import com.clerk.ui.core.common.Spacers
 import com.clerk.ui.core.common.dimens.dp16
 import com.clerk.ui.core.common.dimens.dp8
