+++ source/ui/src/main/java/com/clerk/ui/signup/code/SignUpCodeView.kt
@@ -11,13 +11,13 @@ import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.clerk.ui.R
+import com.clerk.ui.auth.AuthStateEffects
 import com.clerk.ui.auth.LocalAuthState
-import com.clerk.ui.core.common.AuthStateEffects
+import com.clerk.ui.auth.VerificationUiState
+import com.clerk.ui.auth.verificationState
 import com.clerk.ui.core.common.Spacers
-import com.clerk.ui.core.common.VerificationUiState
 import com.clerk.ui.core.common.dimens.dp28
 import com.clerk.ui.core.common.scaffold.ClerkThemedAuthScaffold
-import com.clerk.ui.core.common.verificationState
 import com.clerk.ui.core.input.ClerkCodeInputField
 import com.clerk.ui.core.progress.ClerkLinearProgressIndicator
 import com.clerk.ui.theme.ClerkMaterialTheme
