+++ source/ui/src/main/java/com/clerk/ui/signup/collectfield/SignUpCollectFieldView.kt
@@ -18,13 +18,13 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.clerk.api.Clerk
 import com.clerk.ui.R
+import com.clerk.ui.auth.AuthStateEffects
+import com.clerk.ui.auth.AuthenticationViewState
 import com.clerk.ui.auth.LocalAuthState
 import com.clerk.ui.auth.PreviewAuthStateProvider
 import com.clerk.ui.core.button.standard.ClerkButton
 import com.clerk.ui.core.button.standard.ClerkButtonDefaults
 import com.clerk.ui.core.button.standard.ClerkTextButton
-import com.clerk.ui.core.common.AuthStateEffects
-import com.clerk.ui.core.common.AuthenticationViewState
 import com.clerk.ui.core.common.dimens.dp24
 import com.clerk.ui.core.common.scaffold.ClerkThemedAuthScaffold
 import com.clerk.ui.core.input.ClerkPhoneNumberField
